backend/requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==25.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers
cryptography==45.0.4
defusedxml==0.7.1
Django==4.2.23
django-cors-headers==4.3.1
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.16.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.10
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.3.1
packaging==25.0
Pillow==9.0.0
pluggy==1.0.0.dev0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.6.1
sqlparse==0.5.3
toml==0.10.2
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
